+++ Yandex.Checkout.V3/Amount.cs
@@ -2,7 +2,7 @@
 {
     public class Amount
     {
-        public float value { get; set; }
+        public decimal value { get; set; }
         public string currency { get; set; }
     }
 }
